Create Password.py
The code is a simple password verification program with a limit on the number of attempts. It has a predefined correct password "12345" and allows the user to enter the password up to 3 times. During each attempt:

The program prompts the user to enter the password.
If the entered password matches the correct one, it displays "Correct password!" and breaks the loop using break.
If the entered password is incorrect, it shows "Incorrect password!" and increments the tentativas variable by 1.
If the number of attempts reaches 3, it displays "Number of attempts exceeded."
This code is useful for simulating basic access control with a limited number of attempts before blocking the user.